requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-rw-serializers==1.0.3
drf-url-filters==0.5.1
six==1.14.0
PyYAML==5.3.1
django-extensions==2.2.9

